Collapse duplicate code and convert worklets to real code
